Restore missing temporary variable, deleted by accident in r330845. This unbreaks the ppc32 AIM build.
Reported by:	jhibbits


git-svn-id: svn+ssh://svn.freebsd.org/base/head@330870 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/netinet6/sctp6_var.h
@@ -1,4 +1,6 @@
 /*-
+ * SPDX-License-Identifier: BSD-3-Clause
+ *
  * Copyright (c) 2001-2007, by Cisco Systems, Inc. All rights reserved.
  * Copyright (c) 2008-2012, by Randall Stewart. All rights reserved.
  * Copyright (c) 2008-2012, by Michael Tuexen. All rights reserved.
@@ -49,7 +51,6 @@ sctp6_output(struct sctp_inpcb *, struct mbuf *, struct sockaddr *,
 void sctp6_ctlinput(int, struct sockaddr *, void *);
 void 
 sctp6_notify(struct sctp_inpcb *, struct sctp_tcb *, struct sctp_nets *,
-    uint8_t, uint8_t, uint16_t);
-
+    uint8_t, uint8_t, uint32_t);
 #endif
 #endif
